[Map] Move functions for map endpoints into a separate module
